# prova3OpenAPIDoc.yaml
openapi: 3.0.0
info:
  title: AdaProva3
  contact: {}
  version: '1.0'
servers:
  - url: https://localhost:8080
    variables: {}
paths:
  /iniciar:
    post:
      tags:
        - Misc
      summary: Iniciar Partida
      operationId: IniciarPartida
      parameters:
        - name: Accept
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        '200':
          description: OK
          headers:
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                  example: DENY
            Transfer-Encoding:
              content:
                text/plain:
                  schema:
                    type: string
                  example: chunked
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                  example: Mon, 15 Aug 2022 13:30:19 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: keep-alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IniciarPartida'
              example:
                numeroDaRodada: 1
                cartas:
                  - imdbId: tt5690360
                    titulo: Slender Man
                    ano: 0
                    imagem: https://m.media-amazon.com/images/M/MV5BMjE0MzcwMDAyNl5BMl5BanBnXkFtZTgwMzc4ODg0NDM@._V1_UY67_CR0,0,45,67_AL_.jpg
                    hashImdbId: 5690360
                  - imdbId: tt9731386
                    titulo: Rogue Agent
                    ano: 0
                    imagem: https://m.media-amazon.com/images/M/MV5BYjdkZjE4MzUtMzBhYy00ZGIwLWFmYmYtNzI4NzY1ZjRlYzUzXkEyXkFqcGdeQXVyMTE1MzI2NzIz._V1_UY67_CR0,0,45,67_AL_.jpg
                    hashImdbId: 9731386
                imdbEscolhido: null
      deprecated: false
  /status:
    get:
      tags:
        - Misc
      summary: Status da partida
      operationId: Statusdapartida
      parameters:
        - name: Accept
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        '200':
          description: OK
          headers:
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                  example: DENY
            Transfer-Encoding:
              content:
                text/plain:
                  schema:
                    type: string
                  example: chunked
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                  example: Mon, 15 Aug 2022 13:38:37 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: keep-alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statusdapartida'
              example:
                id: 8
                dataInicio: '2022-08-15T14:37:30.089922'
                jogador:
                  nome: user1
                  ultimoAcesso: '2022-08-15T14:38:09.308499'
                  pontuacaoRecord: 0
                erros: 0
                emAndamento: true
                rodadaAtual:
                  numeroDaRodada: 2
                  cartas:
                    - imdbId: tt0046912
                      titulo: Chamada para a Morte
                      ano: 0
                      imagem: https://m.media-amazon.com/images/M/MV5BOWIwODIxYWItZDI4MS00YzhhLWE3MmYtMzlhZDIwOTMzZmE5L2ltYWdlXkEyXkFqcGdeQXVyNjc1NTYyMjg@._V1_UX45_CR0,0,45,67_AL_.jpg
                      hashImdbId: 46912
                    - imdbId: tt0032553
                      titulo: O Grande Ditador
                      ano: 0
                      imagem: https://m.media-amazon.com/images/M/MV5BMmExYWJjNTktNGUyZS00ODhmLTkxYzAtNWIzOGEyMGNiMmUwXkEyXkFqcGdeQXVyNjU0OTQ0OTY@._V1_UY67_CR0,0,45,67_AL_.jpg
                      hashImdbId: 32553
                  imdbEscolhido: null
                motivoTerminoDaPartida: null
                rodadas:
                  - numeroDaRodada: 1
                    cartas:
                      - imdbId: tt0032138
                        titulo: O Feiticeiro de Oz
                        ano: 0
                        imagem: https://m.media-amazon.com/images/M/MV5BNjUyMTc4MDExMV5BMl5BanBnXkFtZTgwNDg0NDIwMjE@._V1_UX45_CR0,0,45,67_AL_.jpg
                        hashImdbId: 32138
                      - imdbId: tt0308208
                        titulo: Balística
                        ano: 0
                        imagem: https://m.media-amazon.com/images/M/MV5BMTkwMDk0OTcyMF5BMl5BanBnXkFtZTYwNzcxMDg5._V1_UY67_CR0,0,45,67_AL_.jpg
                        hashImdbId: 308208
                    imdbEscolhido: tt0032138
                  - numeroDaRodada: 2
                    cartas:
                      - imdbId: tt0046912
                        titulo: Chamada para a Morte
                        ano: 0
                        imagem: https://m.media-amazon.com/images/M/MV5BOWIwODIxYWItZDI4MS00YzhhLWE3MmYtMzlhZDIwOTMzZmE5L2ltYWdlXkEyXkFqcGdeQXVyNjc1NTYyMjg@._V1_UX45_CR0,0,45,67_AL_.jpg
                        hashImdbId: 46912
                      - imdbId: tt0032553
                        titulo: O Grande Ditador
                        ano: 0
                        imagem: https://m.media-amazon.com/images/M/MV5BMmExYWJjNTktNGUyZS00ODhmLTkxYzAtNWIzOGEyMGNiMmUwXkEyXkFqcGdeQXVyNjU0OTQ0OTY@._V1_UY67_CR0,0,45,67_AL_.jpg
                        hashImdbId: 32553
                    imdbEscolhido: null
                score: 1
      deprecated: false
components:
  schemas:
    IniciarPartida:
      title: IniciarPartida
      required:
        - numeroDaRodada
        - cartas
        - imdbEscolhido
      type: object
      properties:
        numeroDaRodada:
          type: integer
          format: int32
        cartas:
          type: array
          items:
            $ref: '#/components/schemas/Carta'
          description: ''
        imdbEscolhido:
          type: string
          nullable: true
      example:
        numeroDaRodada: 1
        cartas:
          - imdbId: tt5690360
            titulo: Slender Man
            ano: 0
            imagem: https://m.media-amazon.com/images/M/MV5BMjE0MzcwMDAyNl5BMl5BanBnXkFtZTgwMzc4ODg0NDM@._V1_UY67_CR0,0,45,67_AL_.jpg
            hashImdbId: 5690360
          - imdbId: tt9731386
            titulo: Rogue Agent
            ano: 0
            imagem: https://m.media-amazon.com/images/M/MV5BYjdkZjE4MzUtMzBhYy00ZGIwLWFmYmYtNzI4NzY1ZjRlYzUzXkEyXkFqcGdeQXVyMTE1MzI2NzIz._V1_UY67_CR0,0,45,67_AL_.jpg
            hashImdbId: 9731386
        imdbEscolhido: null
    Carta:
      title: Carta
      required:
        - imdbId
        - titulo
        - ano
        - imagem
        - hashImdbId
      type: object
      properties:
        imdbId:
          type: string
        titulo:
          type: string
        ano:
          type: integer
          format: int32
        imagem:
          type: string
        hashImdbId:
          type: integer
          format: int32
      example:
        imdbId: tt5690360
        titulo: Slender Man
        ano: 0
        imagem: https://m.media-amazon.com/images/M/MV5BMjE0MzcwMDAyNl5BMl5BanBnXkFtZTgwMzc4ODg0NDM@._V1_UY67_CR0,0,45,67_AL_.jpg
        hashImdbId: 5690360
    Desistir:
      title: Desistir
      required:
        - id
        - dataInicio
        - jogador
        - erros
        - emAndamento
        - rodadaAtual
        - motivoTerminoDaPartida
        - rodadas
        - score
      type: object
      properties:
        id:
          type: integer
          format: int32
        dataInicio:
          type: string
        jogador:
          $ref: '#/components/schemas/Jogador'
        erros:
          type: integer
          format: int32
        emAndamento:
          type: boolean
        rodadaAtual:
          $ref: '#/components/schemas/RodadaAtual'
        motivoTerminoDaPartida:
          type: string
        rodadas:
          type: array
          items:
            $ref: '#/components/schemas/Rodada'
          description: ''
        score:
          type: integer
          format: int32
      example:
        id: 6
        dataInicio: '2022-08-15T14:36:43.093359'
        jogador:
          nome: user1
          ultimoAcesso: '2022-08-15T14:36:43.092361'
          pontuacaoRecord: 0
        erros: 0
        emAndamento: false
        rodadaAtual:
          numeroDaRodada: 1
          cartas:
            - imdbId: tt0107290
              titulo: Parque Jurássico
              ano: 0
              imagem: https://m.media-amazon.com/images/M/MV5BMjM2MDgxMDg0Nl5BMl5BanBnXkFtZTgwNTM2OTM5NDE@._V1_UX45_CR0,0,45,67_AL_.jpg
              hashImdbId: 107290
            - imdbId: tt0089280
              titulo: Os Hobgoblins
              ano: 0
              imagem: https://m.media-amazon.com/images/M/MV5BMTYzNDYxMTY3N15BMl5BanBnXkFtZTgwNjAxMzkwMzE@._V1_UX45_CR0,0,45,67_AL_.jpg
              hashImdbId: 89280
          imdbEscolhido: null
        motivoTerminoDaPartida: Desistiu
        rodadas:
          - numeroDaRodada: 1
            cartas:
              - imdbId: tt0107290
                titulo: Parque Jurássico
                ano: 0
                imagem: https://m.media-amazon.com/images/M/MV5BMjM2MDgxMDg0Nl5BMl5BanBnXkFtZTgwNTM2OTM5NDE@._V1_UX45_CR0,0,45,67_AL_.jpg
                hashImdbId: 107290
              - imdbId: tt0089280
                titulo: Os Hobgoblins
                ano: 0
                imagem: https://m.media-amazon.com/images/M/MV5BMTYzNDYxMTY3N15BMl5BanBnXkFtZTgwNjAxMzkwMzE@._V1_UX45_CR0,0,45,67_AL_.jpg
                hashImdbId: 89280
            imdbEscolhido: null
        score: 0
    Jogador:
      title: Jogador
      required:
        - nome
        - ultimoAcesso
        - pontuacaoRecord
      type: object
      properties:
        nome:
          type: string
        ultimoAcesso:
          type: string
        pontuacaoRecord:
          type: integer
          format: int32
      example:
        nome: user1
        ultimoAcesso: '2022-08-15T14:36:43.092361'
        pontuacaoRecord: 0
    RodadaAtual:
      title: RodadaAtual
      required:
        - numeroDaRodada
        - cartas
        - imdbEscolhido
      type: object
      properties:
        numeroDaRodada:
          type: integer
          format: int32
        cartas:
          type: array
          items:
            $ref: '#/components/schemas/Carta'
          description: ''
        imdbEscolhido:
          type: string
          nullable: true
      example:
        numeroDaRodada: 1
        cartas:
          - imdbId: tt0107290
            titulo: Parque Jurássico
            ano: 0
            imagem: https://m.media-amazon.com/images/M/MV5BMjM2MDgxMDg0Nl5BMl5BanBnXkFtZTgwNTM2OTM5NDE@._V1_UX45_CR0,0,45,67_AL_.jpg
            hashImdbId: 107290
          - imdbId: tt0089280
            titulo: Os Hobgoblins
            ano: 0
            imagem: https://m.media-amazon.com/images/M/MV5BMTYzNDYxMTY3N15BMl5BanBnXkFtZTgwNjAxMzkwMzE@._V1_UX45_CR0,0,45,67_AL_.jpg
            hashImdbId: 89280
        imdbEscolhido: null
    Rodada:
      title: Rodada
      required:
        - numeroDaRodada
        - cartas
        - imdbEscolhido
      type: object
      properties:
        numeroDaRodada:
          type: integer
          format: int32
        cartas:
          type: array
          items:
            $ref: '#/components/schemas/Carta'
          description: ''
        imdbEscolhido:
          type: string
          nullable: true
      example:
        numeroDaRodada: 1
        cartas:
          - imdbId: tt0107290
            titulo: Parque Jurássico
            ano: 0
            imagem: https://m.media-amazon.com/images/M/MV5BMjM2MDgxMDg0Nl5BMl5BanBnXkFtZTgwNTM2OTM5NDE@._V1_UX45_CR0,0,45,67_AL_.jpg
            hashImdbId: 107290
          - imdbId: tt0089280
            titulo: Os Hobgoblins
            ano: 0
            imagem: https://m.media-amazon.com/images/M/MV5BMTYzNDYxMTY3N15BMl5BanBnXkFtZTgwNjAxMzkwMzE@._V1_UX45_CR0,0,45,67_AL_.jpg
            hashImdbId: 89280
        imdbEscolhido: null
    Prximarodada:
      title: Prximarodada
      required:
        - numeroDaRodada
        - cartas
        - imdbEscolhido
      type: object
      properties:
        numeroDaRodada:
          type: integer
          format: int32
        cartas:
          type: array
          items:
            $ref: '#/components/schemas/Carta'
          description: ''
        imdbEscolhido:
          type: string
          nullable: true
      example:
        numeroDaRodada: 2
        cartas:
          - imdbId: tt0032553
            titulo: O Grande Ditador
            ano: 0
            imagem: https://m.media-amazon.com/images/M/MV5BMmExYWJjNTktNGUyZS00ODhmLTkxYzAtNWIzOGEyMGNiMmUwXkEyXkFqcGdeQXVyNjU0OTQ0OTY@._V1_UY67_CR0,0,45,67_AL_.jpg
            hashImdbId: 32553
          - imdbId: tt0046912
            titulo: Chamada para a Morte
            ano: 0
            imagem: https://m.media-amazon.com/images/M/MV5BOWIwODIxYWItZDI4MS00YzhhLWE3MmYtMzlhZDIwOTMzZmE5L2ltYWdlXkEyXkFqcGdeQXVyNjc1NTYyMjg@._V1_UX45_CR0,0,45,67_AL_.jpg
            hashImdbId: 46912
        imdbEscolhido: null
    Statusdapartida:
      title: Statusdapartida
      required:
        - id
        - dataInicio
        - jogador
        - erros
        - emAndamento
        - rodadaAtual
        - motivoTerminoDaPartida
        - rodadas
        - score
      type: object
      properties:
        id:
          type: integer
          format: int32
        dataInicio:
          type: string
        jogador:
          $ref: '#/components/schemas/Jogador'
        erros:
          type: integer
          format: int32
        emAndamento:
          type: boolean
        rodadaAtual:
          $ref: '#/components/schemas/RodadaAtual'
        motivoTerminoDaPartida:
          type: string
          nullable: true
        rodadas:
          type: array
          items:
            $ref: '#/components/schemas/Rodada'
          description: ''
        score:
          type: integer
          format: int32
      example:
        id: 8
        dataInicio: '2022-08-15T14:37:30.089922'
        jogador:
          nome: user1
          ultimoAcesso: '2022-08-15T14:38:09.308499'
          pontuacaoRecord: 0
        erros: 0
        emAndamento: true
        rodadaAtual:
          numeroDaRodada: 2
          cartas:
            - imdbId: tt0046912
              titulo: Chamada para a Morte
              ano: 0
              imagem: https://m.media-amazon.com/images/M/MV5BOWIwODIxYWItZDI4MS00YzhhLWE3MmYtMzlhZDIwOTMzZmE5L2ltYWdlXkEyXkFqcGdeQXVyNjc1NTYyMjg@._V1_UX45_CR0,0,45,67_AL_.jpg
              hashImdbId: 46912
            - imdbId: tt0032553
              titulo: O Grande Ditador
              ano: 0
              imagem: https://m.media-amazon.com/images/M/MV5BMmExYWJjNTktNGUyZS00ODhmLTkxYzAtNWIzOGEyMGNiMmUwXkEyXkFqcGdeQXVyNjU0OTQ0OTY@._V1_UY67_CR0,0,45,67_AL_.jpg
              hashImdbId: 32553
          imdbEscolhido: null
        motivoTerminoDaPartida: null
        rodadas:
          - numeroDaRodada: 1
            cartas:
              - imdbId: tt0032138
                titulo: O Feiticeiro de Oz
                ano: 0
                imagem: https://m.media-amazon.com/images/M/MV5BNjUyMTc4MDExMV5BMl5BanBnXkFtZTgwNDg0NDIwMjE@._V1_UX45_CR0,0,45,67_AL_.jpg
                hashImdbId: 32138
              - imdbId: tt0308208
                titulo: Balística
                ano: 0
                imagem: https://m.media-amazon.com/images/M/MV5BMTkwMDk0OTcyMF5BMl5BanBnXkFtZTYwNzcxMDg5._V1_UY67_CR0,0,45,67_AL_.jpg
                hashImdbId: 308208
            imdbEscolhido: tt0032138
          - numeroDaRodada: 2
            cartas:
              - imdbId: tt0046912
                titulo: Chamada para a Morte
                ano: 0
                imagem: https://m.media-amazon.com/images/M/MV5BOWIwODIxYWItZDI4MS00YzhhLWE3MmYtMzlhZDIwOTMzZmE5L2ltYWdlXkEyXkFqcGdeQXVyNjc1NTYyMjg@._V1_UX45_CR0,0,45,67_AL_.jpg
                hashImdbId: 46912
              - imdbId: tt0032553
                titulo: O Grande Ditador
                ano: 0
                imagem: https://m.media-amazon.com/images/M/MV5BMmExYWJjNTktNGUyZS00ODhmLTkxYzAtNWIzOGEyMGNiMmUwXkEyXkFqcGdeQXVyNjU0OTQ0OTY@._V1_UY67_CR0,0,45,67_AL_.jpg
                hashImdbId: 32553
            imdbEscolhido: null
        score: 1
    Top10jogadore:
      title: Top10jogadore
      required:
        - nome
        - ultimoAcesso
        - pontuacaoRecord
      type: object
      properties:
        nome:
          type: string
        ultimoAcesso:
          type: string
          nullable: true
        pontuacaoRecord:
          oneOf:
            - type: integer
              format: int32
              nullable: true
            - type: string
              nullable: true
            - nullable: true
      example:
        nome: user1
        ultimoAcesso: '2022-08-15T14:38:09.308499'
        pontuacaoRecord: 0
  securitySchemes:
    basic:
      type: http
      scheme: basic
security:
  - basic: []
tags:
  - name: Misc
    description: ''